Rename useSearchQuery to useSearchDetails
Turns out, useHits() is really slow (> 1ms) and we call it in many
places just to get the hit count.

Let's change this context to be a little more generic so we can put more
stuff in it.

Step 1 is to rename the file.
