+++ frontend/templates/product-list/hooks/useSearchDetails.tsx
